formatting and naming changes for Terms and Nonterminals in the EarleyBNF grammar
also adjusted the indexing for states so that \0 is the whole sequence (equivalent to not specifying a postproc) and \1 is the first, \2 second, etc.  This gains consistency between the token-string selection and rule-state selection enumeration for \-notation.  Updated grammar and documentation re: that.

Updated/fixed comments elsewhere and added spacing between comments and rules in many places, to aid readability.